+++ client/src/app/metadata-schema-form/metadata-schema-form.module.ts
@@ -27,7 +27,7 @@ import {BaseInputComponent} from './metadata-field-types/base-input/base-input.c
 import {EnumDropDownComponent} from "./metadata-field-types/enum-drop-down/enum-drop-down.component";
 import {EnumRadioListComponent} from "./metadata-field-types/enum-radio-list/enum-radio-list.component";
 import {EnumRadioInlineComponent} from "./metadata-field-types/enum-radio-inline/enum-radio-inline.component";
-import {FormItemComponent} from "./metadata-field-types/base-input/form-item.component";
+import {FormItemComponent} from "./form-item/form-item.component";
 
 
 @NgModule({
